+++ t/QuickTime.t
@@ -1,7 +1,7 @@
 # Before "make install", this script should be runnable with "make test".
 # After "make install" it should work as "perl t/QuickTime.t".
 
-BEGIN { $| = 1; print "1..5\n"; $Image::ExifTool::noConfig = 1; }
+BEGIN { $| = 1; print "1..5\n"; $Image::ExifTool::configFile = ''; }
 END {print "not ok 1\n" unless $loaded;}
 
 # test 1: Load the module(s)
@@ -37,6 +37,10 @@ my $testnum = 1;
     $exifTool->SetNewValue('Track1:TrackModifyDate' => '2013:11:04 10:32:15');
     foreach $ext (qw(mov m4a)) {
         ++$testnum;
+        unless (eval { require Time::Local }) {
+            print "ok $testnum # skip Requires Time::Local\n";
+            next;
+        }
         my $testfile = "t/${testname}_$testnum.failed";
         unlink $testfile;
         my $rtnVal = $exifTool->WriteInfo("t/images/QuickTime.$ext", $testfile);
